requirement.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==3.2.13
django-environ==0.4.5
idna==3.6
Pillow==10.1.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==7.7.0
typing_extensions==4.8.0
urllib3==2.1.0
django-markdownx==4.0.0
Pygments==2.17.2

